requirements.txt
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.9
gunicorn==20.0.4
opencv-python-headless
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
opencv-python==4.7.0.68
numpy==1.24.2
Pillow==9.4.0
tensorflow==2.11.0
Flask-Cors==3.0.10
